Plugged TCPLibrary

+++ src/http/controllers/status/models/topic_publisher.rs
@@ -3,6 +3,6 @@ use serde::{Deserialize, Serialize};
 #[derive(Serialize, Deserialize, Debug)]
 pub struct TopicPublisherJsonModel {
     #[serde(rename = "sessionId")]
-    pub session_id: i64,
+    pub session_id: i32,
     pub active: u8,
 }
