+++ src/http/controllers/status/models/topic_queue_subscriber.rs
@@ -5,7 +5,7 @@ pub struct TopicQueueSubscriberJsonModel {
     #[serde(rename = "id")]
     pub subscriber_id: i64,
     #[serde(rename = "sessionId")]
-    pub session_id: i64,
+    pub session_id: i32,
     #[serde(rename = "queueId")]
     pub queue_id: String,
     pub active: u8,
